refactor(api): unify report-suspicious revocation with self/admin revoke
report-suspicious did its own bespoke DELETE on oauth_grants + UPDATE on
oauth_refresh_tokens, marked with a stale TODO(lifecycle) comment about
the not-yet-landed soft-delete column. The lifecycle work has since
landed (same PR), so the two paths can converge — but the hard-DELETE
also broke the new soft-delete audit history that revokedAt + revokedBy
+ revokedReason provides. Worse, the route returned 204 regardless of
whether revocation actually succeeded, so the mobile UI showed
"Session revoked" even when the OAuth grant deletion threw.

- Extract revokeUserOauthClient out of routes/lifecycle.ts into a shared
  services/oauthRevocation.ts module so both lifecycle.ts and approvals.ts
  can call it without a routes-to-routes import.
- approvals.ts report-suspicious: replace inline DELETE + UPDATE with a
  call to revokeUserOauthClient. Captures grantsRevoked, refreshTokensRevoked,
  cacheFailures into the audit details; flips audit result to 'failure' when
  revocation throws.
- When revocation throws or cacheFailures > 0, respond 200 with
  { reported, revoked, cacheFailures, warning } so the mobile UI can
  distinguish "session revoked" from "report recorded — revoke manually".
- Drop the now-unused oauthGrants/oauthRefreshTokens imports from
  approvals.ts.
- Tests: replace the inline-DELETE assertions with revokeUserOauthClient
  call assertions; add cases for revocation-throws (200 + warning, revoked
  false) and cacheFailures > 0 (200 + cacheFailures count).
