+++ apps/api/src/routes/lifecycle.ts
@@ -18,10 +18,9 @@ import { approvalRequests } from '../db/schema/approvals';
 import { authMiddleware, requirePermission } from '../middleware/auth';
 import { rateLimiter, getRedis } from '../services';
 import { writeRouteAudit } from '../services/auditEvents';
-import { revokeGrant, revokeJti } from '../oauth/revocationCache';
 import { ERROR_IDS, logOauthError } from '../oauth/log';
-import { ACCESS_TOKEN_TTL_SECONDS } from '../oauth/provider';
 import { captureException } from '../services/sentry';
+import { revokeUserOauthClient } from '../services/oauthRevocation';
 import { resolveUserAuditOrgId } from './auth/helpers';
 import { PERMISSIONS } from '../services/permissions';
 
@@ -308,138 +307,6 @@ lifecycleRoutes.get('/me/oauth-clients', async (c) => {
     })),
   });
 });
-
-interface RevokeUserOauthClientResult {
-  grantsRevoked: number;
-  refreshTokensRevoked: number;
-  /**
-   * Number of Redis revocation-cache writes that failed after the DB
-   * transaction committed. >0 means the DB state is fully correct but
-   * some active JWTs may survive until natural TTL (≤ ACCESS_TOKEN_TTL).
-   * Callers that need to alert on this should propagate to the response
-   * and audit details.
-   */
-  cacheFailures: number;
-}
-
-async function revokeUserOauthClient(
-  userId: string,
-  clientId: string,
-  revokedByUserId: string,
-  reason: string | null
-): Promise<RevokeUserOauthClientResult> {
-  return asSystem(async () => {
-    const now = new Date();
-
-    // Atomic DB phase. Grants UPDATE + refresh-tokens SELECT/UPDATE must
-    // succeed or fail together; a Redis failure during cache invalidation
-    // (handled after commit) must not be able to leave grants revoked
-    // while refresh tokens stay active in the DB.
-    const { updatedGrants, tokens } = await db.transaction(async (tx) => {
-      const updatedGrants = await tx
-        .update(oauthGrants)
-        .set({ revokedAt: now, revokedByUserId, revokedReason: reason })
-        .where(
-          and(
-            eq(oauthGrants.accountId, userId),
-            eq(oauthGrants.clientId, clientId),
-            isNull(oauthGrants.revokedAt)
-          )
-        )
-        .returning({ id: oauthGrants.id });
-
-      const tokens = await tx
-        .select({
-          id: oauthRefreshTokens.id,
-          payload: oauthRefreshTokens.payload,
-          expiresAt: oauthRefreshTokens.expiresAt,
-        })
-        .from(oauthRefreshTokens)
-        .where(
-          and(
-            eq(oauthRefreshTokens.userId, userId),
-            eq(oauthRefreshTokens.clientId, clientId),
-            isNull(oauthRefreshTokens.revokedAt)
-          )
-        );
-
-      if (tokens.length > 0) {
-        await tx
-          .update(oauthRefreshTokens)
-          .set({ revokedAt: now })
-          .where(inArray(oauthRefreshTokens.id, tokens.map((t) => t.id)));
-      }
-
-      return { updatedGrants, tokens };
-    });
-
-    // Cache-revoke every active jti + grant so sibling JWTs are killed
-    // before natural expiry. Best-effort: failures here are logged + sent
-    // to Sentry, surfaced via cacheFailures, and DO NOT throw. The DB
-    // state above is already correct; throwing here would tell the caller
-    // "revoke failed" when in fact only the cache write failed.
-    let cacheFailures = 0;
-    const onCacheFailure = (
-      err: unknown,
-      message: string,
-      context: Record<string, unknown>
-    ): void => {
-      cacheFailures++;
-      logOauthError({
-        errorId: ERROR_IDS.OAUTH_REVOCATION_CACHE_WRITE_FAILED,
-        message,
-        err,
-        context: { ...context, clientId, userId },
-      });
-      captureException(err);
-    };
-
-    const seenGrants = new Set<string>();
-    for (const tok of tokens) {
-      const payload = tok.payload as { jti?: string; grantId?: string } | null;
-      if (payload?.jti) {
-        const ttl = Math.ceil((new Date(tok.expiresAt).getTime() - Date.now()) / 1000);
-        try {
-          await revokeJti(payload.jti, Math.max(ttl, 1));
-        } catch (err) {
-          onCacheFailure(err, 'self-service jti revocation cache write failed', {
-            jti: payload.jti,
-          });
-        }
-      }
-      if (payload?.grantId && !seenGrants.has(payload.grantId)) {
-        seenGrants.add(payload.grantId);
-        try {
-          await revokeGrant(payload.grantId, ACCESS_TOKEN_TTL_SECONDS);
-        } catch (err) {
-          onCacheFailure(err, 'self-service grant revocation cache write failed', {
-            grantId: payload.grantId,
-          });
-        }
-      }
-    }
-
-    // Belt-and-suspenders: also write a grant marker for any grant rows
-    // that didn't have a refresh token (direct authorize flows).
-    for (const grant of updatedGrants) {
-      if (seenGrants.has(grant.id)) continue;
-      seenGrants.add(grant.id);
-      try {
-        await revokeGrant(grant.id, ACCESS_TOKEN_TTL_SECONDS);
-      } catch (err) {
-        onCacheFailure(err, 'self-service grant-only revocation cache write failed', {
-          grantId: grant.id,
-        });
-      }
-    }
-
-    return {
-      grantsRevoked: updatedGrants.length,
-      refreshTokensRevoked: tokens.length,
-      cacheFailures,
-    };
-  });
-}
 
 lifecycleRoutes.post(
   '/me/oauth-clients/:clientId/revoke',
